pyproject.toml
[project]
name = "salon-ai-supervisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.121.0",
    "livekit>=1.0.17",
    "livekit-agents[silero]~=1.2",
    "livekit-plugins-openai>=1.1.7",
    "livekit-plugins-silero>=0.2.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
